Backend/requirements.txt
# ===================================
# BACKEND DEPENDENCIES - ACTUALIZADO
# Sistema de Análisis Financiero con IA
# ===================================

fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
pandas
numpy
openpyxl==3.1.2
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
openai>=1.0.0
xlsxwriter==3.1.9
alembic==1.12.0
pydantic==2.4.2
pydantic[email]==2.4.2